requirements.txt
flask
sklearn
pandas
numpy
gunicorn
json
joblib
mysql.connector